feat: 🎸 Support more google's domains
