+++ .gitignore
@@ -12,3 +12,4 @@ dist/source/
 key/
 vite.config.åts.timestamp-*
 *.map
+TODO
