+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "new-bing-anywhere",
-  "version": "2.1.0",
+  "version": "2.1.1",
   "private": true,
   "description": "New Bing isn't just for Edge anymore. Anywhere you want",
   "homepage": "https://github.com/haozi/New-Bing-Anywhere",
@@ -34,7 +34,7 @@
     "@types/chrome": "^0.0.239",
     "@types/fs-extra": "^11.0.1",
     "@types/node": "^20.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.60.1",
+    "@typescript-eslint/eslint-plugin": "^5.61.0",
     "copy-and-watch": "^0.1.6",
     "esbuild": "^0.18.11",
     "esbuild-plugin-svgr": "^2.0.0",
