+++ scripts/build.ts
@@ -7,6 +7,7 @@ import fs from 'fs-extra'
 import path from 'path'
 import sortPackageJson from 'sort-package-json'
 import pkg from '../package.json'
+import { GOOGLE_DOMAINS } from '../src/universe/constants'
 import staticRules from './static_rules'
 
 const root = path.join(__dirname, '..')
@@ -70,7 +71,17 @@ const buildChromiumBase = async () => {
     web_accessible_resources: [
       {
         resources: ['inject.js', 'app/*', 'images/*'],
-        matches: ['https://www.bing.com/*', 'https://www.google.com/*', 'https://www.google.com.hk/*']
+        matches: [
+          'https://www.bing.com/*',
+          'https://www.baidu.com/*',
+          'https://www.so.com/*',
+          'https://duckduckgo.com/*',
+          'https://www.ecosia.org/*',
+          'https://*.yandex.com/*',
+          'https://search.brave.com/*',
+          'https://search.naver.com/*',
+          ...GOOGLE_DOMAINS.map((google) => `https://www.${google}/*`)
+        ]
       }
     ],
     options_ui: {
@@ -101,9 +112,17 @@ const buildChromiumBase = async () => {
     host_permissions: [
       // 'notifications',
       // '<all_urls>',
-      '*://*.bing.com/*',
-      '*://www.google.com/*',
-      '*://www.google.com.hk/*'
+      'http://*.bing.com/*',
+      'https://*.bing.com/*',
+      'https://www.baidu.com/*',
+      'https://www.so.com/*',
+      'https://duckduckgo.com/*',
+      'https://www.ecosia.org/*',
+      'https://*.yandex.com/*',
+      'https://search.brave.com/*',
+      'https://search.naver.com/*',
+      'https://*.openai.com/*',
+      ...GOOGLE_DOMAINS.map((google) => `https://www.${google}/search?*`)
     ],
     key: 'MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAxbxQeSdmZpNR6r8FWS5Xviv8NIKPEB1+UpOLsRJHnroPCOSvgZG9u5hbI2ZN0I7DRBXLO3NCxrqcYIp2d62YCzOO4nfKSwnGlAPMFSYw7jyHq0ITjfGIWkql2GsiwRr6MAEM2ktGthDV3iBuL2lRIYfcIOdIUOccxT+2FpDSsncQUHKxjFEisMExX/AAMSNy79PqDUu/5lbEo8zWNlWza5mD69QRU3fK5WGjqrS5naGJ46kPSbE5WU3NPOtHjldPgRVMTbrg6X2GGDGKPp3ISoqj/joNKBNqsMMKn5SURjvqzvzAyVup1/j9XFQ5bGnZYnJTIZ5mvR0wWXnlgf7+RQIDAQAB',
     // content_security_policy: {
@@ -114,7 +133,17 @@ const buildChromiumBase = async () => {
 
     content_scripts: [
       {
-        matches: ['https://www.bing.com/*', 'https://www.google.com/search?*', 'https://www.google.com.hk/search?*'],
+        matches: [
+          'https://www.bing.com/*',
+          'https://www.baidu.com/*',
+          'https://www.so.com/*',
+          'https://duckduckgo.com/*',
+          'https://www.ecosia.org/*',
+          'https://*.yandex.com/*',
+          'https://search.brave.com/*',
+          'https://search.naver.com/*',
+          ...GOOGLE_DOMAINS.map((google) => `https://www.${google}/search?*`)
+        ],
         js: ['zepto.min.js', 'content_script.js'],
         run_at: 'document_start'
       }
@@ -130,7 +159,7 @@ const buildChromiumBase = async () => {
     //   default_title: 'New Bing Anywhere'
     // }
   }
-  fs.outputJSONSync(path.join(chromiumDir, 'manifest.json'), sortManifestJSON(manifest))
+  fs.outputJSONSync(path.join(chromiumDir, 'manifest.json'), sortManifestJSON(manifest), isDev ? { spaces: 2 } : undefined)
 
   // https://developer.chrome.com/docs/webstore/i18n/
   ;['en', 'zh_CN', 'zh_TW', 'ru'].forEach((locale) => {
@@ -164,7 +193,8 @@ const buildChromiumCanary = async () => {
       version: `0.${pkg.version}`,
       homepage_url: 'https://github.com/haozi/New-Bing-Anywhere/tree/canary',
       key: 'MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAp2asctK5nmilg+tZyT74rpsgyfAYWl5pRKsoZDMxj97dwu5YMH1AXoE1ItbFCH8ysjWfsPbYfC0fhFcRljCroPxAJoSl73RRX2rFV8g8aSG101QTYTc2tUvw7xPLk0NS9X4bi/zZmlCHmcoxiOaslN8chs3JgOEQSJROu5PrGpahC9SzZh77iQEtOsYR1grEyuRioFi+x+end1X1tMwaJ4/yYTK4jj9PlFnOKDBFYVhGKCHaWkP2Wv4PPabl/nzUo+l/W0B7fkbaSxI8gir42YzA+OJcPQ/H2UMqtROZxqR847uXsAnB5PfPdo4tT5qUfPd16btsbIr9t6YAMMD0mQIDAQAB'
-    })
+    }),
+    isDev ? { spaces: 2 } : undefined
   )
 }
 
@@ -177,7 +207,8 @@ const buildEdge = async () => {
       ...chromeManifest,
       name: `${pkg.extensionName} (Edge)`,
       key: undefined
-    })
+    }),
+    isDev ? { spaces: 2 } : undefined
   )
 
   fs.outputJSONSync(
@@ -219,7 +250,8 @@ const buildFireFox = async () => {
           // id: 'syntaxright@gmail.com'
         }
       }
-    })
+    }),
+    isDev ? { spaces: 2 } : undefined
   )
 
   fs.removeSync(path.join(firefoxDir, 'app'))
