+++ src/universe/constants.ts
@@ -22,3 +22,207 @@ export const ALL_RESOURCE_TYPES = [
   'webtransport',
   'xmlhttprequest'
 ] as chrome.declarativeNetRequest.ResourceType[]
+
+export const GOOGLE_DOMAINS = [
+  'google.com',
+  'google.com.hk',
+  'google.com.tw',
+  'google.co.in',
+  'google.co.jp',
+  'google.de',
+  'google.es',
+  'google.ru',
+  'google.com.sg',
+  'google.co.uk',
+  'google.co.kr',
+  'google.com.vn',
+  'google.ws',
+  'google.com.tr',
+  'google.se',
+  'google.to',
+
+  'google.ac',
+  'google.ad',
+  'google.ae',
+  'google.al',
+  'google.am',
+  'google.as',
+  'google.at',
+  'google.az',
+  'google.ba',
+  'google.be',
+  'google.bf',
+  'google.bg',
+  'google.bi',
+  'google.bj',
+  'google.bs',
+  'google.bt',
+  'google.by',
+  'google.ca',
+  'google.cat',
+  // 'google.cc',
+  'google.cd',
+  'google.cf',
+  'google.cg',
+  'google.ch',
+  'google.ci',
+  'google.cl',
+  'google.cm',
+  // 'google.cn',
+  'google.co.ao',
+  'google.co.bw',
+  'google.co.ck',
+  'google.co.cr',
+  'google.co.id',
+  'google.co.il',
+  'google.co.ke',
+  'google.co.ls',
+  'google.co.ma',
+  'google.co.mz',
+  'google.co.nz',
+  'google.co.th',
+  'google.co.tz',
+  'google.co.ug',
+  'google.co.uz',
+  'google.co.ve',
+  'google.co.vi',
+  'google.co.za',
+  'google.co.zm',
+  'google.co.zw',
+  'google.com.af',
+  'google.com.ag',
+  'google.com.ai',
+  'google.com.ar',
+  'google.com.au',
+  'google.com.bd',
+  'google.com.bh',
+  'google.com.bn',
+  'google.com.bo',
+  'google.com.br',
+  'google.com.bz',
+  'google.com.co',
+  'google.com.cu',
+  'google.com.cy',
+  'google.com.do',
+  'google.com.ec',
+  'google.com.eg',
+  'google.com.et',
+  'google.com.fj',
+  'google.com.gh',
+  'google.com.gi',
+  'google.com.gt',
+  'google.com.jm',
+  'google.com.kh',
+  'google.com.kw',
+  'google.com.lb',
+  'google.com.lc',
+  'google.com.ly',
+  'google.com.mm',
+  'google.com.mt',
+  'google.com.mx',
+  'google.com.my',
+  'google.com.na',
+  'google.com.nf',
+  'google.com.ng',
+  'google.com.ni',
+  'google.com.np',
+  'google.com.om',
+  'google.com.pa',
+  'google.com.pe',
+  'google.com.pg',
+  'google.com.ph',
+  'google.com.pk',
+  'google.com.pr',
+  'google.com.py',
+  'google.com.qa',
+  'google.com.sa',
+  'google.com.sb',
+  'google.com.sl',
+  'google.com.sv',
+  'google.com.tj',
+  'google.com.ua',
+  'google.com.uy',
+  'google.com.vc',
+  'google.cv',
+  'google.cz',
+  'google.dj',
+  'google.dk',
+  'google.dm',
+  'google.dz',
+  'google.ee',
+  'google.fi',
+  'google.fm',
+  'google.fr',
+  'google.ga',
+  'google.ge',
+  'google.gf',
+  'google.gg',
+  'google.gl',
+  'google.gm',
+  'google.gp',
+  'google.gr',
+  'google.gy',
+  'google.hn',
+  'google.hr',
+  'google.ht',
+  'google.hu',
+  'google.ie',
+  'google.im',
+  'google.io',
+  'google.iq',
+  'google.is',
+  'google.it',
+  'google.je',
+  'google.jo',
+  'google.kg',
+  'google.ki',
+  'google.kz',
+  'google.la',
+  'google.li',
+  'google.lk',
+  'google.lt',
+  'google.lu',
+  'google.lv',
+  'google.md',
+  'google.me',
+  'google.mg',
+  'google.mk',
+  'google.ml',
+  'google.mn',
+  'google.ms',
+  'google.mu',
+  'google.mv',
+  'google.mw',
+  'google.ne',
+  'google.nl',
+  'google.no',
+  'google.nr',
+  'google.nu',
+  'google.pl',
+  'google.pn',
+  'google.ps',
+  'google.pt',
+  'google.ro',
+  'google.rs',
+  'google.rw',
+  'google.sc',
+  'google.sh',
+  'google.si',
+  'google.sk',
+  'google.sm',
+  'google.sn',
+  'google.so',
+  'google.sr',
+  'google.st',
+  'google.td',
+  'google.tg',
+  'google.tk',
+  'google.tl',
+  'google.tm',
+  'google.tn',
+  'google.tt',
+  'google.vg',
+  'google.vu'
+]
+
+export const OTHERS = ['duckduckgo.com', 'www.bing.com', 'www.ecosia.org', 'search.brave.com', 'www.baidu.com']
